refactor(cloud): AppState 子结构 + tenant 模块拆分 + catalog REAL→FLOAT8
- AppState 22 字段 → 13 字段 (StripeConfig, S3Config, EdgeConnections 子结构)
- CaStore 新增 tenant_ca_cache (DashMap 永久缓存)
- tenant.rs 1174 行 → 8 个子模块 (auth/account/store/analytics/order/command/billing/audit)
- 提取 services/rpc.rs 共享 call_edge_rpc 函数
- catalog.rs 消除重复 verify_store，简化 RPC 调用
- 修复 PG migration REAL→DOUBLE PRECISION (SQLx f64 兼容)
- 更新 SQLx 离线元数据

+++ crab-cloud/migrations/0002_catalog_normalized.up.sql
@@ -85,7 +85,7 @@ CREATE TABLE catalog_product_specs (
     product_id    BIGINT  NOT NULL REFERENCES catalog_products(id) ON DELETE CASCADE,
     source_id     BIGINT  NOT NULL,  -- edge 本地 spec ID
     name          TEXT    NOT NULL,
-    price         REAL NOT NULL DEFAULT 0.0,
+    price         DOUBLE PRECISION NOT NULL DEFAULT 0.0,
     display_order INTEGER NOT NULL DEFAULT 0,
     is_default    BOOLEAN NOT NULL DEFAULT FALSE,
     is_active     BOOLEAN NOT NULL DEFAULT TRUE,
@@ -128,7 +128,7 @@ CREATE TABLE catalog_attribute_options (
     attribute_id       BIGINT  NOT NULL REFERENCES catalog_attributes(id) ON DELETE CASCADE,
     source_id          BIGINT  NOT NULL,  -- edge 本地 option ID
     name               TEXT    NOT NULL,
-    price_modifier     REAL NOT NULL DEFAULT 0.0,
+    price_modifier     DOUBLE PRECISION NOT NULL DEFAULT 0.0,
     display_order      INTEGER NOT NULL DEFAULT 0,
     is_active          BOOLEAN NOT NULL DEFAULT TRUE,
     receipt_name       TEXT,
@@ -169,7 +169,7 @@ CREATE TABLE catalog_price_rules (
     target_id        BIGINT,
     zone_scope       TEXT NOT NULL DEFAULT 'all',
     adjustment_type  TEXT NOT NULL,
-    adjustment_value REAL NOT NULL,
+    adjustment_value DOUBLE PRECISION NOT NULL,
     is_stackable     BOOLEAN NOT NULL DEFAULT TRUE,
     is_exclusive     BOOLEAN NOT NULL DEFAULT FALSE,
     valid_from       BIGINT,
